% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_htmls.R
\name{download_htmls}
\alias{download_htmls}
\title{Download GAO Report HTMLs}
\usage{
download_htmls(
  links,
  target_directory = getwd(),
  workers = parallel::detectCores() - 2
)
}
\arguments{
\item{links}{A vector of URLs to download.}

\item{target_directory}{A directory where the HTML files will be saved.}

\item{workers}{Number of workers to use for parallel processing (default: all cores - 2).}
}
\value{
Invisible list of download results. Called for side effect of downloading HTML files.
}
\description{
This function downloads HTML files from a list of GAO report URLs, saving them
to a specified directory. It runs in parallel using available cores and provides progress updates.
}
