% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_pdfs.R
\name{download_pdfs}
\alias{download_pdfs}
\title{Download GAO PDF reports}
\usage{
download_pdfs(
  pdf_links,
  front_url = "https://www.gao.gov",
  download_dir = getwd(),
  workers = parallel::detectCores() - 1,
  sleep_time = 1
)
}
\arguments{
\item{pdf_links}{A character vector of PDF URLs to download (without the domain prefix).}

\item{front_url}{The base URL to prepend to each PDF link (default: "https://www.gao.gov").}

\item{download_dir}{The directory where the PDF files will be saved (default: current working directory).}

\item{workers}{The number of workers to use for parallel downloading (default: all cores - 1).}

\item{sleep_time}{Time (in seconds) to pause between downloads to avoid overwhelming the server (default: 1).}
}
\value{
Invisible list of download results. Called for side effect of downloading PDF files.
}
\description{
This function downloads PDF files from the provided URLs in parallel, saving them to a specified directory.
}
