% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gao-package.R
\docType{package}
\name{gao-package}
\alias{gao}
\alias{gao-package}
\title{gao: A Complete Library of GAO Reports and Metadata}
\description{
A complete library of reports and metadata from the U.S. Government
Accountability Office (GAO). Ships with a bundled dataset of all GAO
report URLs and provides functions for downloading reports as PDF or HTML.
}
\details{
This package is not affiliated with or endorsed by the U.S. Government
Accountability Office. All data is obtained from public web pages at
\url{https://www.gao.gov}.
}
\section{System requirements}{

Browsing the bundled dataset, exporting metadata, and extracting text
from PDFs work out of the box. Downloading report files from gao.gov
additionally requires
\href{https://github.com/lexiforest/curl-impersonate}{curl-impersonate},
a variant of curl that gao.gov's servers accept; functions that need it
print installation instructions if it is missing.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://cetialphafive.github.io/gao/}
  \item \url{https://github.com/CetiAlphaFive/gao}
  \item Report bugs at \url{https://github.com/CetiAlphaFive/gao/issues}
}

}
\author{
\strong{Maintainer}: Jack T. Rametta \email{jtrametta@gmail.com} (\href{https://orcid.org/0000-0002-9841-146X}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Jack T. Rametta \email{jtrametta@gmail.com} (\href{https://orcid.org/0000-0002-9841-146X}{ORCID}) [copyright holder]
}

}
\keyword{internal}
