% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_links.R
\name{extract_links}
\alias{extract_links}
\title{Extract GAO Report Links}
\usage{
extract_links(
  base_url = "https://www.gao.gov/reports-testimonies",
  use_multicore = TRUE,
  workers = NULL,
  last_page = NULL,
  save_to_file = TRUE,
  sleep_timer = 0.5,
  output_file = "gao_report_links.csv"
)
}
\arguments{
\item{base_url}{Character. The base URL for GAO reports (defaults to 'https://www.gao.gov/reports-testimonies').}

\item{use_multicore}{Logical. If TRUE, uses multicore processing (defaults to TRUE).}

\item{workers}{Integer. Number of workers for multicore processing. Only used if use_multicore = TRUE (defaults to 2).}

\item{last_page}{Integer. Optional. Specify the last page manually. If not provided, it will be dynamically detected.}

\item{save_to_file}{Logical. If TRUE, saves the scraped links to a CSV file (defaults to TRUE).}

\item{sleep_timer}{Numeric. The number of seconds to sleep between requests (defaults to .5).}

\item{output_file}{Character. The file path to save the links if \code{save_to_file = TRUE} (defaults to 'gao_report_links.csv').}
}
\value{
A character vector of GAO report links.
}
\description{
This function scrapes all report links from the Government Accountability Office (GAO) reports and testimonies webpage.
It can use multicore processing to speed up scraping.
}
\examples{
\dontrun{
links <- extract_links(use_multicore = TRUE, workers = 10)
}
}
