% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_pdf_links.R
\name{extract_pdf_links}
\alias{extract_pdf_links}
\title{Extract PDF links from GAO reports}
\usage{
extract_pdf_links(
  page_links,
  workers = parallel::detectCores() - 2,
  sleep_time = 1
)
}
\arguments{
\item{page_links}{A vector of report page URLs from which to extract PDF links.}

\item{workers}{Number of workers to use for parallel processing (default: all cores - 2).}

\item{sleep_time}{Time (in seconds) to pause between requests to avoid overwhelming the server (default: 1).}
}
\value{
A vector of unique PDF links.
}
\description{
This function extracts and processes the PDF links from GAO report pages. It uses parallel processing to speed up the extraction.
}
