% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gao-package.R
\docType{package}
\name{gao-package}
\alias{gao}
\alias{gao-package}
\title{gao: Harvest Government Accountability Office Reports}
\description{
Automates the harvesting and downloading of reports published by the United States Government Accountability Office (GAO). Provides functions for scraping report page URLs, extracting PDF download links, and downloading report files in PDF or HTML format. Requires 'curl-impersonate' to be installed on the system to bypass TLS fingerprint filtering.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/CetiAlphaFive/gao}
  \item Report bugs at \url{https://github.com/CetiAlphaFive/gao/issues}
}

}
\author{
\strong{Maintainer}: Jack T. Rametta \email{jtrametta@gmail.com} [copyright holder]

}
\keyword{internal}
