% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gao-package.R
\docType{package}
\name{gao-package}
\alias{gao}
\alias{gao-package}
\title{gao: Harvest Government Accountability Office Reports}
\description{
Tools for harvesting reports published by the U.S. Government
Accountability Office (GAO). Ships with a bundled dataset of all GAO
report URLs and provides functions for downloading reports as PDF or HTML.
}
\details{
This package is not affiliated with or endorsed by the U.S. Government
Accountability Office. All data is obtained from public web pages at
\url{https://www.gao.gov}.
}
\section{System requirements}{

Requires \href{https://github.com/lexiforest/curl-impersonate}{curl-impersonate}
to be installed. GAO.gov uses TLS fingerprint filtering that blocks
standard HTTP clients.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/CetiAlphaFive/gao}
  \item Report bugs at \url{https://github.com/CetiAlphaFive/gao/issues}
}

}
\author{
\strong{Maintainer}: Jack T. Rametta \email{jtrametta@gmail.com} [copyright holder]

}
\keyword{internal}
