% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_links.R
\name{gao_links}
\alias{gao_links}
\title{Get Bundled GAO Report Links}
\usage{
gao_links()
}
\value{
A character vector of GAO report URLs.
}
\description{
Returns the character vector of GAO report URLs bundled with the package.
}
\examples{
links <- gao_links()
length(links)
head(links)
}
